Added CSS, HTML, Fonts, and Images
I updated the styles.CSS and index.HTML files to have a card selection area which was one feature my stakeholders requested. I also trialed different images for the background, added an icon for the browser tab, and downloaded a font to use, which I stored in the fonts folder.