Image carousel
In HTML, I added a div element which contained a list of copyright free images which I will style in CSS and make functional in JavaScript. The main function of this track is that the image perspective moves depending on where the user is on the screen, giving the illusion of looking through a window.