Image Track Functionality
The program tracks the position of the mouse when clicked and converts it into a percentage using a virtual slider. This percentage which is added on to any previous percentage is used to scroll the page. The images also translate depending how the much the page is scrolled, creating a parallax effect. I also created an event listener that detects if the image is clicked and redirects the browser to the HTML file of the country.