Added Animation
In JavaScript, I created an intersection observer that detects when text or image elements are visible on screen. All elements with the "hidden" class in HTML, and which are visible are given the "show" class. With this function, I am able to animate all elements with the "show" class with a transition in CSS.