JavaScript Folder
I created a JavaScript folder as I had two separate sets of code. I could not animate the subpages using the main.js file for whatever reason, so I had to create a separate file for the animation of the subpages. Since I did this, I also needed to change the script tag to the JavaScript folder for each subpage in the HTML folder.