Completed New Zealand, Rarotonga, Singapore, and Vietnam Webpages
I have fully completed the second version of the New Zealand, Rarotonga, Singapore, and Vietnam country page. The page includes a navigation bar, a full image background, a card showcase screen of six recommendations of the best places to visit in the country. In each card, I added a description, a title and a picture of the attraction. I also used more images which I uploaded to the images folder.